requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.2.6
playwright==1.52.0
pyee==13.0.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
django
djangorestframework
